+++ src/components/Shop/Common/ProductWrapper/index.tsx
@@ -3,7 +3,7 @@ import theme from '@styles/theme';
 import SideAd from '../SideAd';
 import OrderBar from '../OptionBars/OrderBar';
 import Product from '@components/Common/Product';
-import Pagination from '@components/Shop/Common/Pagination/Pagination';
+import Pagination from '@components/Shop/Common/Pagination';
 import { productSortOptions } from 'src/utils/shop/sortOption';
 import { handleHeight } from 'src/utils/shop/handleHeight';
 import { ProductAPI, ProductWrapperProps } from 'src/types/shop/types';
